Update: OngoingActivity VM but not working | Update: Refactor New Assistance
