--- Serabutan/Features/View/NewAssistance/NATextFieldAndView.swift
@@ -1,12 +1,0 @@
-//
-//  NATextFieldAndView.swift
-//  Serabutan
-//
-//  Created by Dimas Pramudya Satria H on 08/08/21.
-//
-
-import Foundation
-
-extension NewAssistanceVC {
-    
-}
