--- Serabutan/Features/View/NewAssistance/Utils/NAKeyboardObserver.swift
@@ -1,51 +1,0 @@
-//
-//  NewAssistanceExtentsion.swift
-//  Serabutan
-//
-//  Created by Dimas Pramudya Satria H on 04/08/21.
-//
-
-import Foundation
-import UIKit
-
-extension NewAssistanceVC {
-   
-    //MARK: -Keyboard Delegate
-    func initKeyboardObserver(){
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    
-    @objc func keyboardWillShow(notification: NSNotification) {
-        guard let keyboardSize = (notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue)?.cgRectValue
-        else {
-            // if keyboard size is not available for some reason, dont do anything
-            return
-        }
-        let contentInsets = UIEdgeInsets(top: 0.0, left: 0.0, bottom: keyboardSize.height , right: 0.0)
-        myScrollView.contentInset = contentInsets
-        myScrollView.scrollIndicatorInsets = contentInsets
-    }
-    
-    @objc func keyboardWillHide(notification: NSNotification) {
-        let contentInsets = UIEdgeInsets(top: 0.0, left: 0.0, bottom: 0.0, right: 0.0)
-        myScrollView.contentInset = contentInsets
-        myScrollView.scrollIndicatorInsets = contentInsets
-    }
-    
-    //MARK: - ImagePicker Handler
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        guard let image = info[.editedImage] as? UIImage else { return }
-        print(image, "_ image")
-        dismiss(animated: true)
-
-        currImage = image
-        newAssistanceMediaImage.append(currImage!)
-        
-        DispatchQueue.main.async {
-            self.mediaImageCollectionView.reloadData()
-        }
-    }
-    
-    
-}
